Move `SentrySpanClientHttpRequestInterceptor` to Spring module.
