+++ sentry-spring/api/sentry-spring.api
@@ -76,6 +76,12 @@ public class io/sentry/spring/tracing/SentrySpanAdvice : org/aopalliance/interce
 	public fun invoke (Lorg/aopalliance/intercept/MethodInvocation;)Ljava/lang/Object;
 }
 
+public class io/sentry/spring/tracing/SentrySpanClientHttpRequestInterceptor : org/springframework/http/client/ClientHttpRequestInterceptor {
+	public fun <init> (Lio/sentry/IHub;)V
+	public fun createUriTemplateHandler (Lorg/springframework/web/util/UriTemplateHandler;)Lorg/springframework/web/util/UriTemplateHandler;
+	public fun intercept (Lorg/springframework/http/HttpRequest;[BLorg/springframework/http/client/ClientHttpRequestExecution;)Lorg/springframework/http/client/ClientHttpResponse;
+}
+
 public class io/sentry/spring/tracing/SentryTracingFilter : org/springframework/web/filter/OncePerRequestFilter {
 	public fun <init> (Lio/sentry/IHub;Lio/sentry/SentryOptions;Lio/sentry/spring/SentryRequestResolver;)V
 	protected fun doFilterInternal (Ljavax/servlet/http/HttpServletRequest;Ljavax/servlet/http/HttpServletResponse;Ljavax/servlet/FilterChain;)V
